refactor: extrait la logique de validation dans ValidationToggler (#449)
BuildingValidations ne gère plus que l'affichage ; l'appel serveur, le toaster
et le rafraîchissement passent dans ValidationToggler (simple lien).
