+++ components/BuildingValidations.tsx
@@ -1,22 +1,11 @@
 'use client';
 
-import React, { useState } from 'react';
-import { useDispatch } from 'react-redux';
-import Button from '@codegouvfr/react-dsfr/Button';
+import React from 'react';
 
-import { Actions, AppDispatch } from '@/stores/store';
-import {
-  SelectedBuilding,
-  PublicUser,
-  bdgApiUrl,
-} from '@/stores/map/map-slice';
-import { useRNBFetch } from '@/utils/useRNBFetch';
+import { SelectedBuilding, PublicUser } from '@/stores/map/map-slice';
 import { useRNBAuthentication } from '@/utils/useRNBAuthentication';
 import { hasUserValidated } from '@/utils/validations';
-import {
-  toasterError,
-  toasterSuccess,
-} from '@/components/contribution/toaster';
+import ValidationToggler from '@/components/ValidationToggler';
 
 import styles from '@/styles/buildingValidations.module.scss';
 
@@ -29,10 +18,7 @@ export default function BuildingValidations({
   building,
   allowEdit,
 }: BuildingValidationsProps) {
-  const dispatch: AppDispatch = useDispatch();
-  const { fetch } = useRNBFetch();
   const { user } = useRNBAuthentication();
-  const [isLoading, setIsLoading] = useState(false);
 
   const validatedBy = building.validated_by;
   const userHasValidated = hasUserValidated(validatedBy, user?.username);
@@ -40,56 +26,25 @@ export default function BuildingValidations({
   // En consultation, rien à afficher s'il n'y a aucune validation.
   if (!allowEdit && validatedBy.length === 0) return null;
 
-  const setValidation = async (isValid: boolean) => {
-    setIsLoading(true);
-    try {
-      const response = await fetch(bdgApiUrl(`${building.rnb_id}/`), {
-        method: 'PATCH',
-        body: JSON.stringify({ is_valid: isValid }),
-      });
-      if (!response.ok) {
-        toasterError(
-          dispatch,
-          'Erreur lors de la mise à jour de la validation',
-        );
-      } else {
-        toasterSuccess(
-          dispatch,
-          isValid ? 'Validation enregistrée' : 'Validation retirée',
-        );
-        // Re-consultation du bâtiment pour rafraîchir validated_by (réponse 204).
-        await dispatch(Actions.map.selectBuilding(building.rnb_id));
-      }
-    } catch (err: any) {
-      toasterError(
-        dispatch,
-        err.message || 'Erreur lors de la mise à jour de la validation',
-      );
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
   const actionButton = allowEdit && user && (
     <div className={styles.actionShell}>
-      <a href="#" onClick={() => setValidation(!userHasValidated)}>
-        {userHasValidated ? 'Retirer ma validation' : 'Valider ce bâtiment'}
-      </a>
+      <ValidationToggler
+        rnbId={building.rnb_id}
+        label={
+          userHasValidated ? 'Retirer ma validation' : 'Valider ce bâtiment'
+        }
+        isValid={!userHasValidated}
+      />
     </div>
   );
 
   // Cas édition sans aucune validation : bloc neutre + bouton "Valider".
   if (validatedBy.length === 0) {
     return (
-      <div className={styles.section}>
-        <div className={styles.validationEmpty}>
-          <div>
-            <h2 className={styles.sectionTitle}>Bâtiment non validé</h2>
-            <div className={styles.sectionBody}>
-              <span>Aucune validation pour le moment.</span>
-              {actionButton}
-            </div>
-          </div>
+      <div className={styles.validationEmpty}>
+        <div>
+          La géométrie, le statut et les adresses de ce bâtiment vous semblent
+          corrects ? {actionButton}
         </div>
       </div>
     );
